requirements.txt
# Core Django
Django>=4.2.0,<5.0
djangorestframework>=3.14.0

# WebSockets & Real-time
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.0.0

# Firebase Authentication
firebase-admin>=6.0.0

# Database & ORM
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# CORS
django-cors-headers>=4.0.0

# Geolocation & Spatial
geopy>=2.3.0
django-rest-framework-gis>=0.20.0

# Logging & Monitoring
python-json-logger>=2.0.0
django-extensions>=3.2.0

# File Storage
django-storages>=1.14.0  # For S3 and other cloud storage
boto3>=1.26.0  # AWS SDK for S3

# Pagination & Filtering
django-filter>=23.0
drf-flex-fields>=1.0.0

# Authentication & Permissions
PyJWT>=2.6.0
cryptography>=39.0.0

# HTTP Client
requests>=2.28.0

# Task Queue (optional, for async tasks)
celery>=5.2.0
redis>=4.3.0

# API Documentation
drf-spectacular>=0.25.0

# Validation
django-phonenumber-field>=7.0.0
phonenumbers>=8.13.0

# env management
python-decouple>=3.8

# Testing
pytest>=7.2.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Development
django-debug-toolbar>=3.8.0
